fix(units): address PR #175 review feedback
- weather-map EU visibility legend: relabel to the true km equivalents of the
  SM breakpoints visibilityColor() uses (8.0/4.8/1.6 km) so the legend text
  matches the dot colors at the boundaries
- PreferencesUpdate.units_region: validate as Literal["auto","europe","us"]
  (rejects invalid values at the HTTP boundary)
- load_service_toggles: annotate -> dict[str, Any] (returns str + bool values)
- airport_conditions: float(round(...)) to match visibility_m: float|None
- packs: comment that mixed/transatlantic routes resolve auto -> europe
- advisories-ui: comment why formatCondVis uses != null (catches old-pack undefined)

+++ src/weatherbrief/analysis/airport_conditions.py
@@ -199,7 +199,7 @@ def _compute_for_airport(
 
         if hourly:
             if hourly.visibility_m is not None:
-                visibility_m = round(hourly.visibility_m)
+                visibility_m = float(round(hourly.visibility_m))
                 visibility_sm = round(hourly.visibility_m / _M_PER_SM, 1)
             wind_speed_kt = hourly.wind_speed_10m_kt
             wind_direction_deg = hourly.wind_direction_10m_deg
